+++ pkgs/artisanal/example/tui/examples/exec/main.dart
@@ -3,7 +3,7 @@ library;
 
 import 'dart:io' as io;
 
-import 'package:artisanal/tui.dart' as tui;
+import 'package:artisanal/runtime.dart' as tui;
 
 class EditorFinishedMsg extends tui.Msg {
   const EditorFinishedMsg(this.result);